// frontend/src/api/generated/models/EmailMessageHeader.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Todo Microservice API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 *
 * @export
 * @interface EmailMessageHeader
 */
export interface EmailMessageHeader {
  /**
   *
   * @type {string}
   * @memberof EmailMessageHeader
   */
  from?: string;
  /**
   *
   * @type {string}
   * @memberof EmailMessageHeader
   */
  subject?: string;
  /**
   *
   * @type {Date}
   * @memberof EmailMessageHeader
   */
  date?: Date;
}

/**
 * Check if a given object implements the EmailMessageHeader interface.
 */
export function instanceOfEmailMessageHeader(value: object): value is EmailMessageHeader {
  return true;
}

export function EmailMessageHeaderFromJSON(json: any): EmailMessageHeader {
  return EmailMessageHeaderFromJSONTyped(json, false);
}

export function EmailMessageHeaderFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean
): EmailMessageHeader {
  if (json == null) {
    return json;
  }
  return {
    from: json['from'] == null ? undefined : json['from'],
    subject: json['subject'] == null ? undefined : json['subject'],
    date: json['date'] == null ? undefined : new Date(json['date']),
  };
}

export function EmailMessageHeaderToJSON(json: any): EmailMessageHeader {
  return EmailMessageHeaderToJSONTyped(json, false);
}

export function EmailMessageHeaderToJSONTyped(
  value?: EmailMessageHeader | null,
  ignoreDiscriminator: boolean = false
): any {
  if (value == null) {
    return value;
  }

  return {
    from: value['from'],
    subject: value['subject'],
    date: value['date'] == null ? undefined : value['date'].toISOString(),
  };
}